add user profile activities for crated thread and replies

+++ app/Http/Controllers/ThreadsController.php
@@ -61,8 +61,17 @@ class ThreadsController extends Controller
 
     public function destroy($channel,Thread $thread)
     {
+        $this->authorize('update',$thread);
+        if($thread->user_id != auth()->user()->id)
+        {
+            abort(403,'You do not have permission to do this');
+        }
         $thread->delete();
-        return response([],204);
+        if( \request()->wantsJson())
+        {
+           return response([],204);
+        }
+        return redirect('/threads');
     }
 
     /**
